Some work to remove /W4 warnings from Tinker, but ultimately the changes aren't qualitatively improving the code, so I stopped. We'll stick with /W3 in general.

+++ tinker/glgui/rootpanel.cpp
@@ -54,7 +54,7 @@ CRootPanel::CRootPanel() :
 	m_iMX = 0;
 	m_iMY = 0;
 
-	m_iQuad = ~0;
+	m_iQuad = (size_t)~0;
 }
 
 CRootPanel::~CRootPanel( )
